chore: convert the app to use SceneDelegate
